第一次提交

+++ backend/src/main/kotlin/lan/jing/backend/controller/TradeController.kt
@@ -1,7 +1,9 @@
 package lan.jing.backend.controller
 
 import com.mybatisflex.core.query.QueryWrapper
+import com.mybatisflex.kotlin.extensions.db.filterOne
 import com.mybatisflex.kotlin.extensions.db.insert
+import com.mybatisflex.kotlin.extensions.kproperty.eq
 import lan.jing.backend.entity.MedicinePurchase
 import lan.jing.backend.entity.MedicineRestock
 import lan.jing.backend.mapper.PurchaseMapper
@@ -9,6 +11,8 @@ import lan.jing.backend.mapper.RestockMapper
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.data.redis.core.RedisTemplate
 import org.springframework.web.bind.annotation.*
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
 import java.util.*
 
 @RestController
@@ -16,7 +20,8 @@ import java.util.*
 class TradeController {
     data class PurchaseRequest(
         val rid: Long,
-        val count: Int
+        val count: Int,
+        val time: String
     )
 
     data class PurchaseResponse(
@@ -43,21 +48,23 @@ class TradeController {
         else {
             val restock =
                 restockMapper.selectOneWithRelationsById(purchaseRequest.rid) ?: return PurchaseResponse("", "购买失败")
-            if (restock.count!! < purchaseRequest.count) return PurchaseResponse("", "购买失败")
+            if (restock.presentCount!! < purchaseRequest.count) return PurchaseResponse("", "购买失败")
+            val localDate = LocalDate.parse(purchaseRequest.time, DateTimeFormatter.ofPattern("yyyy-MM-dd"))
+            val localDateTime = localDate.atStartOfDay()
             insert(
                 MedicinePurchase(
                     mid = restock.mid,
                     uid = userId.toLong(),
                     rid = purchaseRequest.rid,
                     count = purchaseRequest.count,
-                    price = restock.price,
-                    time = Date()
+                    price = restock.outPrice,
+                    time = localDateTime
                 )
             )
             restockMapper.update(
                 MedicineRestock(
                     id = purchaseRequest.rid,
-                    count = restock.count!! - purchaseRequest.count
+                    presentCount = restock.presentCount!! - purchaseRequest.count
                 )
             )
             return PurchaseResponse("200", "购买成功")
@@ -67,7 +74,9 @@ class TradeController {
     data class RestockRequest(
         val mid: Long,
         val count: Int,
-        val price: Double
+        val price: Double,
+        val outPrice: Double,
+        val time: String
     )
 
     data class RestockResponse(
@@ -83,13 +92,17 @@ class TradeController {
         val userId = redisTemplate.opsForValue().get("user:$token")
         if (userId == null) return RestockResponse("", "进货失败")
         else {
+            val localDate = LocalDate.parse(restockRequest.time, DateTimeFormatter.ofPattern("yyyy-MM-dd"))
+            val localDateTime = localDate.atStartOfDay()
             insert(
                 MedicineRestock(
                     mid = restockRequest.mid,
                     uid = userId.toLong(),
                     count = restockRequest.count,
+                    presentCount = restockRequest.count,
                     price = restockRequest.price,
-                    time = Date()
+                    outPrice = restockRequest.outPrice,
+                    time = localDateTime
                 )
             )
             return RestockResponse("200", "进货成功")
@@ -99,7 +112,8 @@ class TradeController {
     data class RestockEditRequest(
         val id: Long,
         val count: Int,
-        val price: Double
+        val price: Double,
+        val time: String
     )
 
     data class RestockEditResponse(
@@ -112,11 +126,18 @@ class TradeController {
     fun editRestock(
         @RequestBody restockEditRequest: RestockEditRequest
     ): RestockEditResponse {
+        val restock = filterOne<MedicineRestock> {
+            MedicineRestock::id eq restockEditRequest.id
+        } ?: return RestockEditResponse("400", "进货失败")
+        if (restock.count!! < restockEditRequest.count) return RestockEditResponse("400", "进货失败")
+        val localDate = LocalDate.parse(restockEditRequest.time, DateTimeFormatter.ofPattern("yyyy-MM-dd"))
+        val localDateTime = localDate.atStartOfDay()
         restockMapper.update(
             MedicineRestock(
                 id = restockEditRequest.id,
-                count = restockEditRequest.count,
-                price = restockEditRequest.price
+                presentCount = restockEditRequest.count,
+                outPrice = restockEditRequest.price,
+                time = localDateTime
             )
         )
         return RestockEditResponse("200", "进货成功")
@@ -167,11 +188,90 @@ class TradeController {
 
     @PostMapping("/deleteRestock")
     fun deleteRestock(@RequestBody deletePurchaseRequest: DeletePurchaseRequest): RestockResponse {
+        if (deletePurchaseRequest.id < 16) return RestockResponse("400", "默认进货不可删除")
         try {
             restockMapper.deleteById(deletePurchaseRequest.id)
         } catch (e: Exception) {
             return RestockResponse("400", "删除失败")
         }
         return RestockResponse("200", "删除成功")
+    }
+
+    data class ResStockCountByYearResponse(
+        val code: String,
+        val message: String,
+        val counts: List<Int>,
+        val years: List<String>
+    )
+
+    @GetMapping("/restockCountByYear")
+    fun stockCountByYear(): ResStockCountByYearResponse {
+        //近七年
+        val years = mutableListOf<String>()
+        val counts = mutableListOf<Int>()
+        val year = Calendar.getInstance().get(Calendar.YEAR)
+        for (i in 0..6) {
+            years.add((year - i).toString())
+            counts.add(restockMapper.countByYear(year - i))
+        }
+        return ResStockCountByYearResponse("200", "获取成功", counts, years)
+    }
+
+    data class ResStockCountByMonthResponse(
+        val code: String,
+        val message: String,
+        val counts: List<Int>,
+        val months: List<String>
+    )
+
+    @GetMapping("restockCountByMonth")
+    fun stockCountByMonth(): ResStockCountByMonthResponse {
+        //近七个月
+        val months = mutableListOf<String>()
+        val counts = mutableListOf<Int>()
+        val year = Calendar.getInstance().get(Calendar.YEAR)
+        val month = Calendar.getInstance().get(Calendar.MONTH)
+        for (i in 0..6) {
+            if (month - i < 0) {
+                months.add("${year - 1}-${month - i + 12}")
+                counts.add(restockMapper.countByMonth(year - 1, month - i + 12))
+            } else {
+                months.add("$year-${month - i + 1}")
+                counts.add(restockMapper.countByMonth(year, month - i + 1))
+            }
+        }
+        return ResStockCountByMonthResponse("200", "获取成功", counts, months)
+    }
+
+    data class ResStockCountByDayResponse(
+        val code: String,
+        val message: String,
+        val counts: List<Int>,
+        val days: List<String>
+    )
+
+    @GetMapping("restockCountByDay")
+    fun stockCountByDay(): ResStockCountByDayResponse {
+        //近七天
+        val days = mutableListOf<String>()
+        val counts = mutableListOf<Int>()
+        val year = Calendar.getInstance().get(Calendar.YEAR)
+        val month = Calendar.getInstance().get(Calendar.MONTH)
+        val day = Calendar.getInstance().get(Calendar.DAY_OF_MONTH)
+        for (i in 0..6) {
+            if (day - i < 0) {
+                if (month - 1 < 0) {
+                    days.add("${year - 1}-${month + 11}-${day - i + 30}")
+                    counts.add(restockMapper.countByDay(year - 1, month + 11, day - i + 30))
+                } else {
+                    days.add("$year-${month + 1}-${day - i + 30}")
+                    counts.add(restockMapper.countByDay(year, month + 1, day - i + 30))
+                }
+            } else {
+                days.add("$year-${month + 1}-${day - i}")
+                counts.add(restockMapper.countByDay(year, month + 1, day - i))
+            }
+        }
+        return ResStockCountByDayResponse("200", "获取成功", counts, days)
     }
 }
